refactor(agent-task): extract runner-exec identity from lifecycle_ops (#9927)
